Move RFID to dedicated app and add QR sidebar
